+++ src/services/location.service.ts
@@ -1,9 +1,12 @@
-import { GPScompare } from '@/dtos/request/location.dto';
+import { GPScompare, GetCoordinatesDto } from '@/dtos/request/location.dto';
 import { GetLocationResponse } from '@/dtos/response/location.response';
 import { ErrorEnum } from '@/exceptions/errorCodes';
 import { HttpException } from '@/exceptions/httpException';
 import { LocationModel } from '@/models/location.model';
+import { ResidentialInfoModel } from '@/models/residentialInfo.model';
+import { ResidentialPeerModel } from '@/models/residentialPeer.model';
 import { Geolocation } from '@/remote/location/location';
+import { residentialPeerService } from './residentialPeer.service';
 
 class LocationService {
   public async createLocation(userId: string, address: string): Promise<GetLocationResponse> {
@@ -54,6 +57,41 @@ class LocationService {
     } catch (err) {
       throw new HttpException(ErrorEnum.INVALID_COORDINATES);
     }
+  }
+
+  public async capturePeerLocation(peerUUID: string, data: GetCoordinatesDto) {
+    const { peerId } = await residentialPeerService.peerUUIDtoPeerId(peerUUID);
+    const peer = await ResidentialPeerModel.findById(peerId);
+    if (!peer) {
+      throw new HttpException(ErrorEnum.PEER_NOT_FOUND);
+    }
+    const residentialInfo = await ResidentialInfoModel.findById(peer.ref);
+    if (!residentialInfo) {
+      throw new HttpException(ErrorEnum.RESIDENTIAL_INFO_NOT_FOUND);
+    }
+    const location = await LocationModel.create({
+      user: peer.user,
+      latitude: data.latitude,
+      longitude: data.longitude,
+    });
+    residentialInfo.capturedLocation = location._id;
+    residentialInfo.save();
+    return { success: true, message: 'Location Captured' };
+  }
+
+  public async captureUserLocation(userId: string, data: GetCoordinatesDto) {
+    const residentialInfo = await ResidentialInfoModel.findOne({ user: userId });
+    if (!residentialInfo) {
+      throw new HttpException(ErrorEnum.RESIDENTIAL_INFO_NOT_FOUND);
+    }
+    const location = await LocationModel.create({
+      user: userId,
+      latitude: data.latitude,
+      longitude: data.longitude,
+    });
+    residentialInfo.capturedLocation = location._id;
+    residentialInfo.save();
+    return residentialInfo;
   }
 }
